Simplification of type checking type specifications
When type checking a type declaration we would need to add the "self"
variable to the local environment where its type is its model.  This
commit simplifies creating the type of the model in the case where
there are multiple model fields and the model type is a record with
each of those fields.