Refactor type checking model declarations
When processing the model field of a type declaration, we add a record
type to the global environment in the there are case of named fields.
This was done in the same function where model declarations are type
checked.  This commit splits updating the environment and type
checking the model into two different functions.