Instantiation of aliases and models is done by Solver
In the namespace module, we had a function that instantiated that type
variables in case of a type alias.  Since the namespace module should
only concern itself with mapping names to the their relevant
information, this logic was moved to the Solver module.  Additionally,
we use the constraints mechanism to instantiate the type.