Model is built from the default lens
Before, there was only one model associated to one type.  Whenever we
wanted the model for a type, we would perform a lookup based on the
name of the type.  Now we lookup what the default lens and then
combine it with the OCaml type to generate its model.  Gospel is not
made more expressive with this commit, but the infrastructure for
explicit lenses has been laid out.

Note: Although type variables are syntactically allowed in model
definitions they are not supported.